refactor: Step2 리뷰반영 - view 관련 부분 빼기
